requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-braces==1.15.0
django-model-utils==4.2.0
django-phonenumber-field==6.0.0
django-star-ratings==0.9.1
django-widget-tweaks==1.4.9
phonenumbers==8.12.37
phonenumberslite==8.12.37
Pillow==8.4.0
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
swapper==1.2.0
